+++ nix/.stack.nix/cardano-sl-util.nix
@@ -104,6 +104,7 @@
             (hsPkgs.time-units)
             (hsPkgs.universum)
             (hsPkgs.unordered-containers)
+            (hsPkgs.yaml)
           ];
           build-tools = [
             (hsPkgs.buildPackages.hspec-discover)